requirements.txt
streamlit
requests
python-dotenv
matplotlib
seaborn
